Finger Count 2022-10-08 03-14-57.mp4
This is a video that showcases the work of the project. Had to keep it short because of time constraints. When the hand is completely closed there is confusion because the code uses a circle that it draws based on the farthest point in the polygon drawn around the contour of the hand. The circle drawn has an area that is small due to which even folded fingers lie outside it. But as the fingers are opened it becomes accurate.